BREAKING CHANGE: Coprl presenter version 3 update.
Version 3 of presenters has a breaking change around the way that plugins work.
It also renames the toplevel namespace to 'Coprl'.
This commit used the `coprl generator plugin foo` command to rebuild the plugin to be compatitble with version 3.
