+++ foo_presenter_plugin.gemspec
@@ -4,11 +4,11 @@ $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 Gem::Specification.new do |spec|
   spec.name          = "foo_presenter_plugin"
   spec.version       = '0.0.1'
-  spec.authors       = ["Russell Edens"]
-  spec.email         = ["rx@voomify.com"]
+  spec.authors       = ["Rx"]
+  spec.email         = ["russell@russelledens.net"]
 
-  spec.summary       = %q{Simplest example of a presenter plugin.}
-  spec.homepage      = 'http://github.com/rx/foo_presenters_plugin'
+  spec.summary       = %q{foo presenter plugin.}
+  spec.homepage      = 'http://github.com/Rx/foo_presenters_plugin'
   spec.license       = "MIT"
 
   spec.files         = `git ls-files -z`.split("\x0").reject do |f|
@@ -17,5 +17,5 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_development_dependency "bundler", "~> 1.16"
-  spec.add_development_dependency "rake", "~> 12.3"
+  spec.add_development_dependency "rake", "~> 10.0"
 end
