+++ lib/coprl/presenters/plugins/foo.rb
@@ -1,0 +1,9 @@
+# Loads the plugin features
+# Helpers
+require_relative 'foo/helpers/foo_helper'
+# Components
+require_relative 'foo/components/foo_dsl'
+require_relative 'foo/web_client/components/foo_render'
+# Actions
+require_relative 'foo/components/actions/foo_dsl'
+require_relative 'foo/web_client/components/actions/foo_data'
