--- lib/voom/presenters/plugins/foo/foo_header.erb
@@ -1,20 +1,0 @@
-<style>
-  <%= File.read(File.expand_path('foo.css', __dir__)  ) %>
-</style>
-
-<script>
-  // This file is defining the callback method that will be invoked when the custom plugin action `bar` is encountered
-  <%= File.read(File.expand_path('bar.js', __dir__)  ) %>
-  // This file is defining the plugin component that will be delegated to by the WebClient.
-  <%= File.read(File.expand_path('foo.js', __dir__)  ) %>
-
-  // Custom javascript specific to this sample plugin - sets the class name that causes the initial load to
-  // animate a transition (defined in the foo.css file)
-  function load() {
-    document.getElementById('random_fact').className = ' load';
-  }
-
-  window.onload = load;
-</script>
-
-
