+++ views/assets/js/components/foo.js
