+++ views/components/_foo.erb
@@ -9,6 +9,6 @@ hookup the event handling markup
 
 Any additional data you want to pass to your component javascript should be setup here as data attributes.
 -->
-<iframe id="<%= comp.id %>" class='v-plugin v-input' data-plugin-callback='RandomFacts'
+<iframe id="random_fact" class='v-plugin v-input' data-plugin-callback='RandomFacts'
         src="<%= comp.random_fact %>" height=512
-        <%= erb :"components/event", :locals => {comp: comp, events: comp.events, parent_id: comp.id} %>></iframe>
+        <%= partial "components/event", :locals => {comp: comp, events: comp.events, parent_id: comp.id} %>></iframe>
