+++ views/components/foo.erb
@@ -1,0 +1,14 @@
+<!--Custom markup.
+If the plugin defines a component then it should set v-plugin and data-plugin-callback to the class of the component
+
+If the plugin posts data then it should define the v-input class.
+That will cause its validate and prepareSubmit methods in the component to be called
+
+If the plugin has events it publishes then the erb :"components/event" needs to be invoked to
+hookup the event handling markup
+
+Any additional data you want to pass to your component javascript should be setup here as data attributes.
+-->
+<iframe id="random_fact" class='v-plugin v-input' data-plugin-callback='RandomFacts'
+        src="<%= comp.random_fact %>" height=512
+        <%= erb :"components/event", :locals => {comp: comp, events: comp.events, parent_id: comp.id} %>></iframe>
